+++ src/api/card.ts
@@ -1,49 +1,81 @@
 import client from "./axios";
 
-export const createComment = async (
-  content: string,
-  cardId: number,
-  columnId: number,
-  dashboardId: number
-) => {
-  const { data } = await client.post("comments", {
-    content: content,
-    cardId: cardId,
-    columnId: columnId,
-    dashboardId: dashboardId,
+export const getCards = async (id: string) => {
+  const { data } = await client.get("cards", {
+    params: { columnId: id, size: 100 },
   });
-
-  return data;
+  return data.cards;
 };
 
-export const getComments = async (
-  size = 10,
-  cursorId: number | null,
+export const changeCard = async (id: string, body: object) => {
+  try {
+    const res = await client.put(`cards/${id}`, body);
+    return res;
+    // console.log(res);
+  } catch (err) {
+    console.log(err);
+  }
+};
+
+export const createCard = async (
+  cardData: any,
+  dashboardId: number,
+  columnId: number
+) => {
+  try {
+    const requestData: any = {
+      assigneeUserId: parseInt(cardData.assigneeUserId),
+      dashboardId: dashboardId,
+      columnId: columnId,
+      title: cardData.title,
+      description: cardData.description,
+      dueDate: cardData.dueDate,
+      tags: [...cardData.tags],
+    };
+
+    // imageUrl 값이 존재하고 string 타입인 경우에만 imageUrl 속성을 추가합니다.
+    if (cardData.imageUrl && typeof cardData.imageUrl === "string") {
+      requestData.imageUrl = cardData.imageUrl;
+    }
+    const res = await client.post(`cards/`, requestData);
+    return res;
+  } catch (err) {
+    console.log(err);
+  }
+};
+
+export const deleteCard = async (cardId: number) => {
+  try {
+    const res = await client.delete(`cards/${cardId}`);
+    return res;
+  } catch (e: any) {
+    console.log(e);
+  }
+};
+
+export const updateCard = async (
+  cardData: any,
   columnId: number,
   cardId: number
 ) => {
-  const { data } = await client.get("comments", {
-    params: {
-      size: size,
-      cursorId: cursorId,
+  try {
+    const requestData: any = {
+      assigneeUserId: parseInt(cardData.assigneeUserId),
       columnId: columnId,
-      cardId: cardId,
-    },
-  });
-
-  return data;
-};
-
-export const deleteComment = async (commentId: number) => {
-  const { data } = await client.delete(`/comments/${commentId}`);
-
-  return data;
-};
-
-export const updateComment = async (commentId: number, content: string) => {
-  const { data } = await client.put(`/comments/${commentId}`, {
-    content: content,
-  });
-
-  return data;
+      title: cardData.title,
+      description: cardData.description,
+      dueDate: cardData.dueDate,
+      tags: [...cardData.tags],
+    };
+    // imageUrl 값이 존재하고 string 타입인 경우에만 imageUrl 속성을 추가합니다.
+    if (cardData.imageUrl && typeof cardData.imageUrl === "string") {
+      requestData.imageUrl = cardData.imageUrl;
+    }
+    const res = await client.put(`cards/${cardId}`, {
+      ...requestData,
+    });
+    return res;
+  } catch (e: any) {
+    console.log(e);
+  }
 };
