+++ src/pages/DashBoardPage/DashBoardPage.tsx
@@ -1,7 +1,6 @@
 // @ts-nocheck
 import { DragDropContext } from "react-beautiful-dnd";
 import { useQueryClient, useMutation } from "@tanstack/react-query";
-import { changeCard } from "../../api/dashboard";
 import styles from "./DashBoardPage.module.css";
 import Columns from "../../components/Column/Columns";
 import { useModal } from "../../hooks/useModal";
@@ -9,10 +8,10 @@ import { createPortal } from "react-dom";
 import ColumnCreateModal from "../../components/Modal/ColumnCreateModal";
 import { useNavigate, useParams } from "react-router-dom";
 import { useQueries, useQuery } from "@tanstack/react-query";
-import { getCards, getColumns } from "../../api/dashboard";
+import { getColumns } from "../../api/dashboard";
+import { getCards, changeCard } from "../../api/card";
 import { ColumnType } from "../../interface/DashboardType";
 import BaseButton from "../../components/BaseButton/BaseButton";
-import AddIcon from "../../assets/addButton.svg";
 
 export default function DashBoardPage() {
   const { isModalOpen, handleModalOpen, handleModalClose } = useModal();
